feat(mcp): move MCP implementation to strands.mcp, deprecate old path
Moves the MCP implementation from strands.tools.mcp to strands.mcp using
git-tracked renames. strands.tools.mcp becomes a deprecation alias:

- Emits a DeprecationWarning on import pointing to strands.mcp.
- Re-exports the public API (object identity preserved).
- Registers legacy submodule paths (strands.tools.mcp.mcp_client, etc.)
  via sys.modules so ``from strands.tools.mcp.mcp_client import X``
  continues to resolve to the canonical modules.

Follow-up PR will move the tests into tests/strands/mcp, update
integration tests, README, and AGENTS.md.

Refs #1431
