+++ strands-py/src/strands/mcp/__init__.py
@@ -1,17 +1,14 @@
-"""Canonical import path for the Model Context Protocol (MCP) integration.
+"""Model Context Protocol (MCP) integration.
 
-Model Context Protocol functionality is being promoted from
-``strands.tools.mcp`` to this top-level ``strands.mcp`` package because MCP
-now spans tools, prompts, resources, tasks, and elicitation -- concepts
-that extend beyond ``tools``.
+This package provides integration with the Model Context Protocol (MCP), allowing agents to use tools provided by MCP
+servers.
 
-For now this package is a thin re-export of ``strands.tools.mcp``. A
-follow-up change will invert the relationship: the implementation will
-live here and ``strands.tools.mcp`` will become a deprecated alias.
-Users can safely migrate imports to ``strands.mcp`` today; the public
-API is identical and object identity is preserved.
+- Docs: https://www.anthropic.com/news/model-context-protocol
 """
 
-from ..tools.mcp import MCPAgentTool, MCPClient, MCPTransport, TasksConfig, ToolFilters
+from .mcp_agent_tool import MCPAgentTool
+from .mcp_client import MCPClient, ToolFilters
+from .mcp_tasks import TasksConfig
+from .mcp_types import MCPTransport
 
 __all__ = ["MCPAgentTool", "MCPClient", "MCPTransport", "TasksConfig", "ToolFilters"]
