+++ strands-py/src/strands/mcp/mcp_client.py
@@ -44,11 +44,11 @@ from mcp.types import TextContent as MCPTextContent
 from pydantic import AnyUrl
 from typing_extensions import Protocol, TypedDict
 
-from ...types import PaginatedList
-from ...types.exceptions import MCPClientInitializationError, ToolProviderException
-from ...types.media import ImageFormat
-from ...types.tools import AgentTool, ToolResultContent, ToolResultStatus
-from ..tool_provider import ToolProvider
+from ..tools.tool_provider import ToolProvider
+from ..types import PaginatedList
+from ..types.exceptions import MCPClientInitializationError, ToolProviderException
+from ..types.media import ImageFormat
+from ..types.tools import AgentTool, ToolResultContent, ToolResultStatus
 from .mcp_agent_tool import MCPAgentTool
 from .mcp_instrumentation import mcp_instrumentation
 from .mcp_tasks import DEFAULT_TASK_CONFIG, DEFAULT_TASK_POLL_TIMEOUT, DEFAULT_TASK_TTL, TasksConfig
